Create a volume and set the metadata
